Cleanup of the header mess: removed the redundant APIC/mp structure definitions from smp_priv.h and stage2_priv.h. Created a new file smp_apic.h which now contains those definitions. Made the APIC_xxx register offsets only base offsets (without referencing a particular variable). Other small cleanups (i.e. "unsigned int" to uint32).
git-svn-id: file:///srv/svn/repos/haiku/trunk/current@2604 a95241bf-73f2-0310-859d-f6bbb57e9c96

+++ headers/private/kernel/stage2_struct.h
@@ -5,6 +5,8 @@
 #ifndef _NEWOS_BOOT_STAGE2_STRUCT_H
 #define _NEWOS_BOOT_STAGE2_STRUCT_H
 
+#include <SupportDefs.h>
+
 // must match SMP_MAX_CPUS in arch_smp.h
 #define MAX_BOOT_CPUS 4
 #define MAX_PHYS_MEM_ADDR_RANGE 4
@@ -12,8 +14,8 @@
 #define MAX_PHYS_ALLOC_ADDR_RANGE 4
 
 typedef struct ar {
-	unsigned long start;
-	unsigned long size;
+	uint32 start;
+	uint32 size;
 } addr_range;
 
 #endif
